initial repo: pre-copy
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@11 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ l4/pkg/demangle/lib/src/memcmp.c
@@ -38,10 +38,12 @@
  *	contents are identical upto the length of s1.
  */
 
-int memcmp(const void *s1v, const void *s2v, unsigned size);
+#include <stddef.h>
+
+int memcmp(const void *s1v, const void *s2v, size_t size);
 
 int
-memcmp(const void *s1v, const void *s2v, unsigned size)
+memcmp(const void *s1v, const void *s2v, size_t size)
 {
 	register const char *s1 = s1v, *s2 = s2v;
 	register unsigned int a, b;
